Move poster and art to a mixin
