+++ plexapi/library.py
@@ -4,6 +4,7 @@ from urllib.parse import quote, quote_plus, unquote, urlencode
 from plexapi import X_PLEX_CONTAINER_SIZE, log, media, utils
 from plexapi.base import OPERATORS, PlexObject, PlexPartialObject
 from plexapi.exceptions import BadRequest, NotFound
+from plexapi.mixins import PosterArt
 from plexapi.settings import Setting
 from plexapi.utils import deprecated
 
@@ -1525,7 +1526,7 @@ class FirstCharacter(PlexObject):
 
 
 @utils.registerPlexObject
-class Collections(PlexPartialObject):
+class Collections(PlexPartialObject, PosterArt):
     """ Represents a single Collection.
 
         Attributes:
