+++ plexapi/playlist.py
@@ -5,12 +5,13 @@ from plexapi import utils
 from plexapi.base import Playable, PlexPartialObject
 from plexapi.exceptions import BadRequest, NotFound, Unsupported
 from plexapi.library import LibrarySection
+from plexapi.mixins import PosterArt
 from plexapi.playqueue import PlayQueue
 from plexapi.utils import cast, toDatetime
 
 
 @utils.registerPlexObject
-class Playlist(PlexPartialObject, Playable):
+class Playlist(PlexPartialObject, Playable, PosterArt):
     """ Represents a single Playlist.
 
         Attributes:
